refactor(setup): split 767-line route.ts into 10 focused modules under lib/setup/
Extract telegram, github, seed-db, cron, gateway, workspace, rolodex, email,
smoke, and constants modules. Route.ts is now a 105-line orchestrator that
imports and calls each step. No functionality change — POST response shape
is identical.
